requirements.txt
torch
sentence-transformers
faiss-cpu 
pymongo
python-dotenv
